pyproject.toml
[project]
name = "localcowork-lite"
version = "0.1.0"
description = "On-device AI agent with MCP tool-calling - runs entirely on RTX 3060 6GB"
requires-python = ">=3.11"

dependencies = [
    # FastAPI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    # MCP
    "mcp>=1.0.0",
    # LLM client (OpenAI-compat for llama.cpp)
    "openai>=1.50.0",
    "httpx>=0.27.0",
    # Knowledge / RAG
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    # Document processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pillow>=10.0.0",
    # Utilities
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
